Refactor Time Picker To Be More Code Readable
Refactor the Time Picker code to be more readable because it has
become to complex with having the formula inside its popup. Which will
make testing both the TimePicker and popup hard to do.

So the plan for the Time Picker is as follows. Split the time when it
comes into the Time Picker, from Frig, and send into a new class
called Time Formatter. Where the time will be formatted for the
Time Picker Popup to use.

If the user changes the time in the Time Picker Popup the time will
be combine and sent back to the Time Formatter, in Time Picker,
before sending to the Time Picker and to Frig in general.