Remove Layout As Required In Label
Remove the layout as a required label prop because in test the layout
may not be supply as it comes from the Unbound Input. So in order to
keep the test simple and decoupled from Unbound Input it the layout
should be on its own. Furthermore, the default value is vertical.