Move Value Link Select Code From Frig Core
Move the Value Linked Select code from Frig Core to Frigging-Bootstrap's
select because Value Link is no longer put of React. Also, Value Link
was unable to access the value in React DOM's select, Value Linked
Select fixed this. With the removal of Value Link, from React, having
this fix is no longer needed.

This change will also allow Frig theme developers to create themes with
their own version of select and not have to use the default select in
Frig core allow for a smaller code base in the core.