Update React Component to Use Latest ESLint Rules
Apply the Following ESLint Rules:

 * Update React Component Refs because Refs string is going to be
eventually deprecate. Refs strings are considered legacy and will be
remove in the future at some point.

As a result of this findDOMNode is going away as well from
eslint-plugin-react. Since there is not real reason to use findDOMNode
without a Ref.

Refer:

https://facebook.github.io/react/docs/more-about-refs.html#the-ref-string-attribute
https://github.com/yannickcr/eslint-plugin-react/issues/678#issue-165177220
https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/no-string-refs.md
https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/no-find-dom-node.md

 * Add Declared Global Variables To Components That Use Them in order
for the developer to know which variables will be accessible in the
global space.

Refer:

http://eslint.org/docs/rules/no-undef

 * Move the import to have the NPM packages to be first for
eslint-plugin-import. Because this will make understandable which
packages and files are being added to the component for the developer

Refer:

https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/imports-first.md

 * Fix mix different operation on the same line because this will
make it hard to clarify which operation belongs to which statement.

Refer:

http://eslint.org/docs/rules/no-mixed-operators

 * Turn off JSX filename extension requirement for ESlint React rule
because while it is understandable to know which files are using JSX
as it is not part of the JavaScript standard. The issue either or not
to have JSX filename extension has not been settled yet in the
community. So we will allow both JS and JSX extensions for now.

Refer:

https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/jsx-filename-extension.md

 * Fix Not Using Self Closing Tags When Warranted

Refer:

https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/self-closing-comp.md

 * Add which extraneous packages to use in ESlint rule because by
default all extraneous package in package.json dependencies,
devDependencies, optionalDependencies and peerDependencies are
forbid to be import.

This will forces the developer to declare which package.json
extraneous they are planning on use and not accidentally using a
extraneous they may not intended to use.

Refer:

https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-extraneous-dependencies.md

 * Remove the Continue Statement from sizeClassNames in Util because
the Continue Statement makes it harder to read and maintain.

Refer:

http://eslint.org/docs/rules/no-continue

 * Fix Not Using Shorthand Operator, in TimePicker Popup

Refer:

http://eslint.org/docs/rules/operator-assignment

 * Add htmlFor to the Label component because this will id the form
control the label is attach to for accessible of the form.

Refer:

https://github.com/evcohen/eslint-plugin-jsx-a11y/blob/master/docs/rules/label-has-for.md

 * Add const to the inputLabel & saved in Input Component because
those do not change in their function.

Refer:

http://eslint.org/docs/rules/prefer-const